refactor(cp-mzn): rename differential-linear model and tests to mzn_ prefix
