+++ claasp/cipher_modules/models/cp/mzn_models/mzn_differential_linear_model.py
